Fix database loading when using as an ASDF module
ASDF compiles its source files and stores them in a special cache
folder, so that they are no longer sitting among the rest of the system.
Thus, we need a way of getting the original folder location for the ASDF
system on the compiled FASL file. For that to happen, we use a
macro (which gets executed at compile-time), saving the compile-time
path of the file, so that we can find the static files.